Refactored dataset metadata views location in source in preparation for showing it on PAP.
