+++ projects/laji/src/app/shared-modules/dataset-metadata/dataset-metadata-viewer/dataset-metadata-viewer-item/dataset-metadata-viewer-item.component.ts
