+++ projects/laji/src/app/shared-modules/dataset-metadata/dataset-metadata-viewer/dataset-metadata-viewer.component.scss
